+++ hashes/src/ripemd160.rs
@@ -85,7 +85,7 @@ impl crate::HashEngine for HashEngine {
 
     fn n_bytes_hashed(&self) -> u64 { self.bytes_hashed }
 
-    engine_input_impl!();
+    crate::internal_macros::engine_input_impl!();
 }
 
 #[cfg(feature = "small-hash")]
